requirements.txt
kafka-python==2.0.2
pyarrow==14.0.2
pandas==2.0.3